refactor: centralize battery type constants, improve MAC normalization and widget update handling
