+++ app/src/beta/res/values-de/strings.xml
@@ -1,0 +1,4 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <string name="app_name">clOwOck (Beta)</string>
+</resources>
