+++ app/src/beta/res/values-zh/strings.xml
@@ -1,0 +1,4 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <string name="app_name">clOwOck（测试版）</string>
+</resources>
