+++ app/src/main/java/com/mrboombastic/buwudzik/data/BatteryType.kt
@@ -1,0 +1,7 @@
+package com.mrboombastic.buwudzik.data
+
+/** Stored in [DeviceProfile.batteryType] and QR [DeviceShareData.batteryType]. */
+object BatteryType {
+    const val ALKALINE = "alkaline"
+    const val NIMH = "nimh"
+}
